add some testing for Body component & children
